Separate out keystore from wallets in monorepo and split up files (#8170)
* Move keystores out of wallets sdk directory

* Split keystores into separate files

* Update dependency graph

* Fix config paths
