+++ packages/sdk/keystores/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@celo/wallet-keystore",
+  "name": "@celo/keystores",
   "version": "1.2.2-dev",
   "description": "keystore implementation",
   "author": "Celo",
@@ -14,7 +14,7 @@
   "scripts": {
     "build": "tsc -b .",
     "clean": "tsc -b . --clean",
-    "docs": "typedoc && ts-node ../../utils/scripts/linkdocs.ts wallet-keystore",
+    "docs": "typedoc && ts-node ../utils/scripts/linkdocs.ts keystores",
     "test": "jest --runInBand",
     "lint": "tslint -c tslint.json --project .",
     "prepublishOnly": "yarn build"
