+++ packages/sdk/keystores/src/keystore-base.ts
@@ -4,8 +4,6 @@ import {
   privateKeyToAddress,
 } from '@celo/utils/lib/address'
 import Wallet from 'ethereumjs-wallet'
-import { mkdirSync, promises as fsPromises, readFileSync, unlinkSync, writeFileSync } from 'fs'
-import path from 'path'
 
 export enum ErrorMessages {
   KEYSTORE_ENTRY_EXISTS = 'Existing encrypted keystore for address',
@@ -146,82 +144,5 @@ export abstract class KeystoreBase {
    */
   async deleteKeystore(address: string) {
     this.removeKeystore(await this.getKeystoreName(address))
-  }
-}
-
-export class InMemoryKeystore extends KeystoreBase {
-  /**
-   * Used for mocking keystore operations in unit tests
-   */
-  private _storage: Record<string, string> = {}
-
-  // Implements required abstract class methods
-  persistKeystore(keystoreName: string, keystore: string) {
-    this._storage[keystoreName] = keystore
-  }
-
-  getRawKeystore(keystoreName: string): string {
-    return this._storage[keystoreName]
-  }
-
-  async getAllKeystoreNames(): Promise<string[]> {
-    return Object.keys(this._storage)
-  }
-
-  removeKeystore(keystoreName: string) {
-    delete this._storage[keystoreName]
-  }
-}
-
-export class FileKeystore extends KeystoreBase {
-  /**
-   * Implements keystore as a directory on disk
-   * with files for keystore entries
-   */
-  private _keystoreDir: string
-
-  /**
-   * Creates (but does not overwrite existing) directory
-   * for containing keystore entries.
-   * @param keystoreDir Path to directory where keystore will be saved
-   */
-  constructor(keystoreDir: string) {
-    super()
-    this._keystoreDir = path.join(keystoreDir, 'keystore')
-    // Does not overwrite existing directories
-    mkdirSync(this._keystoreDir, { recursive: true })
-  }
-
-  /**
-   * @returns List of file names (keystore entries) in the keystore
-   */
-  async getAllKeystoreNames(): Promise<string[]> {
-    return fsPromises.readdir(this._keystoreDir)
-  }
-
-  /**
-   * Saves keystore entries as a file in the keystore directory
-   * @param keystoreName File name of keystore entry
-   * @param keystore V3Keystore string entry
-   */
-  persistKeystore(keystoreName: string, keystore: string) {
-    writeFileSync(path.join(this._keystoreDir, keystoreName), keystore)
-  }
-
-  /**
-   * Gets contents of keystore entry file
-   * @param keystoreName File name of keystore entry
-   * @returns V3Keystore string entry
-   */
-  getRawKeystore(keystoreName: string): string {
-    return readFileSync(path.join(this._keystoreDir, keystoreName)).toString()
-  }
-
-  /**
-   * Deletes file keystore entry from directory
-   * @param keystoreName File name of keystore entry to be removed
-   */
-  removeKeystore(keystoreName: string) {
-    return unlinkSync(path.join(this._keystoreDir, keystoreName))
   }
 }
