+++ packages/sdk/keystores/tsconfig.json
@@ -4,6 +4,5 @@
     "rootDir": "src",
     "outDir": "lib"
   },
-  "include": ["src"],
-  "references": [{ "path": "../../utils" }]
+  "include": ["src/**/*", "types/**/*"]
 }
