--- packages/sdk/wallets/wallet-keystore/src/index.ts
@@ -1,2 +1,0 @@
-export * from './keystore-wallet-wrapper'
-export * from './keystores'
